pyproject.toml
[project]
name = "attendance-system"
version = "0.1.0"
description = "A project to implement an attendance system using face recognition"
readme = "README.md"
requires-python = "==3.10.17"
dependencies = [
    "ai-edge-litert>=1.3.0",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "opencv-python>=4.11.0.86",
    "peewee>=3.18.1",
    "pyside6>=6.9.0",
    "pygame>=2.5.0",
    "requests>=2.32.3",
    "tensorflow",
    "gunicorn>=23.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.uv]
package = true

[project.scripts]
attendance-app = "src.main:main"
flask-app = "src.api.test:app"
